partial cleanup usings - move MicroCHP data structs into correct header
