+++ src/EnergyPlus/MicroCHPElectricGenerator.hh
@@ -62,6 +62,184 @@ namespace MicroCHPElectricGenerator {
     extern Array1D_bool CheckEquipName;
     extern Array1D_bool MySizeFlag;
 
+    struct MicroCHPParamsNonNormalized
+    {
+        // Members
+        // user parameters
+        std::string Name;         // name of this PowerModule data
+        Real64 MaxElecPower;      // net electric power [W]
+        Real64 MinElecPower;      // net electric power [W]
+        Real64 MinWaterMdot;      // minimum cooling water flow [kg/s]
+        Real64 MaxWaterTemp;      // limit temp for inlet cooling water [C]
+        int ElecEffCurveID;       // index for TriQuadratic for electrical efficiency
+        int ThermalEffCurveID;    // index for TriQuadric for thermal efficiency
+        bool InternalFlowControl; // Plant or Internal Flow rate control?
+        bool PlantFlowControl;    // default is plant control
+        int WaterFlowCurveID;     // index for BiQuadratic for water flow rate internal control
+        int AirFlowCurveID;       // index for Quadratic for generator air flow
+        Real64 DeltaPelMax;       // max rate of change in net electric power [W/s}
+        Real64 DeltaFuelMdotMax;  // Maximum Rate of change in fuel flow rate [kmol/s2]
+        Real64 UAhx;              // heat exchanger UA [W/K]
+        Real64 UAskin;            // skin loss UA [W/K]
+        Real64 RadiativeFraction; // skin loss fraction to radiant energy []
+        Real64 MCeng;             // aggregated thermal mass of engine [J/K]
+        Real64 MCcw;              // aggregated thermal mass of heat recovery [J/k]
+        Real64 Pstandby;          // standby power [w]
+        bool WarmUpByTimeDelay;   // Warm up mode control
+        bool WarmUpByEngineTemp;  // Warm up mode control
+        Real64 kf;                // coefficient k_f for warmup fuel flow rate
+        Real64 TnomEngOp;         // nominal engine operating temperature [C]
+        Real64 kp;                // coefficient k_p for warmup power
+        Real64 Rfuelwarmup;       // Warm Up Fuel Flow Rate Limit Ratio
+        Real64 WarmUpDelay;       // time for warm up delay [s]
+        Real64 PcoolDown;         // power during cool down
+        Real64 CoolDownDelay;     // time for cool down delay [s]
+        bool MandatoryFullCoolDown;
+        bool WarmRestartOkay;
+        // calculated and from elsewhere
+        Real64 TimeElapsed; // Fraction of the current hour that has elapsed (h)
+        // Saved in order to identify the beginning of a new system time
+        int OpMode;
+        Real64 OffModeTime;      // amount of time generator spent in Off mode
+        Real64 StandyByModeTime; // amount of time generator spent in standby mode
+        Real64 WarmUpModeTime;   // amount of time generator spent in warm up mode
+        Real64 NormalModeTime;   // amount of time generator spent in normal mode
+        Real64 CoolDownModeTime; // amount of time generator spent in Cool down mode
+        Real64 TengLast;         // last timestep's value for engine temperature
+        Real64 TempCWOutLast;    // last timestep's value for cooling water outlet temperature
+        Real64 Pnet;
+        Real64 ElecEff;
+        Real64 Qgross;
+        Real64 ThermEff;
+        Real64 Qgenss;
+        Real64 NdotFuel;
+        Real64 MdotFuel;
+        Real64 Teng;
+        Real64 TcwIn;
+        Real64 TcwOut;
+        Real64 MdotAir;
+        Real64 QdotSkin; // rate of heat loss to zone
+        Real64 QdotConvZone;
+        Real64 QdotRadZone;
+
+        // Default Constructor
+        MicroCHPParamsNonNormalized()
+                : MaxElecPower(0.0), MinElecPower(0.0), MinWaterMdot(0.0), MaxWaterTemp(0.0), ElecEffCurveID(0), ThermalEffCurveID(0),
+                  InternalFlowControl(false), PlantFlowControl(true), WaterFlowCurveID(0), AirFlowCurveID(0), DeltaPelMax(0.0), DeltaFuelMdotMax(0.0),
+                  UAhx(0.0), UAskin(0.0), RadiativeFraction(0.0), MCeng(0.0), MCcw(0.0), Pstandby(0.0), WarmUpByTimeDelay(false),
+                  WarmUpByEngineTemp(true), kf(0.0), TnomEngOp(0.0), kp(0.0), Rfuelwarmup(0.0), WarmUpDelay(0.0), PcoolDown(0.0), CoolDownDelay(0.0),
+                  MandatoryFullCoolDown(false), WarmRestartOkay(true), TimeElapsed(0.0), OpMode(0), OffModeTime(0.0), StandyByModeTime(0.0),
+                  WarmUpModeTime(0.0), NormalModeTime(0.0), CoolDownModeTime(0.0), TengLast(20.0), TempCWOutLast(20.0), Pnet(0.0), ElecEff(0.0),
+                  Qgross(0.0), ThermEff(0.0), Qgenss(0.0), NdotFuel(0.0), MdotFuel(0.0), Teng(20.0), TcwIn(20.0), TcwOut(20.0), MdotAir(0.0),
+                  QdotSkin(0.0), QdotConvZone(0.0), QdotRadZone(0.0)
+        {
+        }
+    };
+
+    struct MicroCHPReportDataStruct // these are all for reporting only!
+    {
+        // Members
+        int Mode;                    // report operating mode (dev only, remove at end)
+        Real64 OffModeTime;          // amount of time generator spent in Off mode
+        Real64 StandyByModeTime;     // amount of time generator spent in standby mode
+        Real64 WarmUpModeTime;       // amount of time generator spent in warm up mode
+        Real64 NormalModeTime;       // amount of time generator spent in normal mode
+        Real64 CoolDownModeTime;     // amount of time generator spent in Cool down mode
+        Real64 ACPowerGen;           // reporting: power (W)
+        Real64 ACEnergyGen;          // reporting: energy (J)
+        Real64 QdotGross;            // reporting: interim gross power (W)
+        Real64 Qgenss;               // reporting: net recovered heat rate steadystate(0)
+        Real64 QdotHX;               // reporting: rate of heat exchange from engine to coolant (W)
+        Real64 QdotHR;               // reporting: rate of heat recovered (W)
+        Real64 Tengine;              // reporting: engine mass temperature (C)
+        Real64 TotalHeatEnergyRec;   // reporting: total heat recovered (J)
+        Real64 ExhaustEnergyRec;     // reporting: exhaust gas heat recovered (J)
+        Real64 FuelEnergyLHV;        // reporting: Fuel Energy used in Lower Heating Value(J)
+        Real64 FuelEnergyUseRateLHV; // reporting: Fuel Energy used in Lower Heating Value(W)
+        Real64 FuelEnergyHHV;        // reporting: Fuel Energy used in Higher Heating Value(J)
+        Real64 FuelEnergyUseRateHHV; // reporting: Fuel Energy used in Higher Heating Value(W)
+        Real64 HeatRecInletTemp;     // reporting: Heat Recovery Loop Inlet Temperature (C)
+        Real64 HeatRecOutletTemp;    // reporting: Heat Recovery Loop Outlet Temperature (C)
+        Real64 HeatRecMdot;          // reporting: Heat Recovery Loop Mass flow rate (kg/s)
+        // air supply and blower
+        Real64 TairInlet; // State point 1
+        Real64 MdotAir;   // air flow in kmol/sec
+        // fuel supply and compressor
+        Real64 TfuelInlet;           // State point 2 [C]
+        Real64 NdotFuel;             // fuel flow in [kmol/sec]
+        Real64 MdotFuel;             // fuel flow in [kg/s]
+        Real64 FuelCompressPower;    // electrical power used by fuel supply compressor [W]
+        Real64 FuelCompressEnergy;   // electrical energy used by fuel supply compressor [J]
+        Real64 FuelCompressSkinLoss; // heat rate of losses.by fuel supply compressor [W]
+        // heat exchanger for water to exhaust heat recovery
+        //   REAL(r64)    :: qHX = 0.0d0 ! heat flow from gas stream to water [W]
+        //   REAL(r64)    :: HXenergy = 0.0d0 !energy from gas stream to water [J]
+        //   REAL(r64)    :: THXexh = 0.0d0 ! temperature of exhaust gases leaving heat exchanger.
+        //   REAL(r64)    :: WaterVaporFractExh = 0.0d0 ! water vapor fraction in exhaust gas stream
+        // relative to water vapor entering HX  (NdotH20/Ndoaux-mix)
+        //    INTEGER :: SeqSubstIterations = 0 ! number of iterations in SOFC loop
+        //    INTEGER :: RegulaFalsiIterations = 0 ! number of iterations in Tproduct gas solving
+        Real64 SkinLossPower;   // heat loss to surrounding zone [W]
+        Real64 SkinLossEnergy;  // heat loss to surround zone [J]
+        Real64 SkinLossConvect; // convective heat loss to zone [W]
+        Real64 SkinLossRadiat;  // radiative heat loss to zone [W}
+        Real64 ElectEfficiency;
+        Real64 ThermalEfficiency;
+        Real64 OverallEfficiency;
+
+        // Default Constructor
+        MicroCHPReportDataStruct()
+                : Mode(0), OffModeTime(0.0), StandyByModeTime(0.0), WarmUpModeTime(0.0), NormalModeTime(0.0), CoolDownModeTime(0.0), ACPowerGen(0.0),
+                  ACEnergyGen(0.0), QdotGross(0.0), Qgenss(0.0), QdotHX(0.0), QdotHR(0.0), Tengine(0.0), TotalHeatEnergyRec(0.0), ExhaustEnergyRec(0.0),
+                  FuelEnergyLHV(0.0), FuelEnergyUseRateLHV(0.0), FuelEnergyHHV(0.0), FuelEnergyUseRateHHV(0.0), HeatRecInletTemp(0.0),
+                  HeatRecOutletTemp(0.0), HeatRecMdot(0.0), TairInlet(0.0), MdotAir(0.0), TfuelInlet(0.0), NdotFuel(0.0), MdotFuel(0.0),
+                  FuelCompressPower(0.0), FuelCompressEnergy(0.0), FuelCompressSkinLoss(0.0), SkinLossPower(0.0), SkinLossEnergy(0.0),
+                  SkinLossConvect(0.0), SkinLossRadiat(0.0), ElectEfficiency(0.0), ThermalEfficiency(0.0), OverallEfficiency(0.0)
+        {
+        }
+    };
+
+    struct MicroCHPDataStruct
+    {
+        // Members
+        // user input data
+        std::string Name;                     // name of this Micro CHP Generator
+        std::string ParamObjName;             // name of parameter object
+        MicroCHPParamsNonNormalized A42Model; // Nested parameter data structure
+        bool ModelTypeAnnex42;                // normalized =  non-normalized?
+        Real64 NomEff;                        // nominal efficiency
+        std::string ZoneName;
+        int ZoneID;
+        std::string PlantInletNodeName;
+        int PlantInletNodeID;
+        std::string PlantOutletNodeName;
+        int PlantOutletNodeID;
+        Real64 PlantMassFlowRate;              // only if internal control
+        Real64 PlantMassFlowRateMax;           // hardware limit for node%massflowrateMax
+        bool PlantMassFlowRateMaxWasAutoSized; // true if mass flow rate was autosized on input
+        std::string AirInletNodeName;
+        int AirInletNodeID;
+        std::string AirOutletNodeName;
+        int AirOutletNodeID;
+        MicroCHPReportDataStruct Report; // structure of report variables
+        int FuelSupplyID;                // index for fuel supply data structure
+        int DynamicsControlID;           // index in GeneratorDynamics data where control issues are handled
+        int AvailabilitySchedID;         // index for availability schedule
+        int CWLoopNum;                   // cooling water plant loop index number
+        int CWLoopSideNum;               // cooling water plant loop side index
+        int CWBranchNum;                 // cooling water plant loop branch index
+        int CWCompNum;                   // cooling water plant loop component index
+
+        // Default Constructor
+        MicroCHPDataStruct()
+                : ModelTypeAnnex42(true), NomEff(0.0), ZoneID(0), PlantInletNodeID(0), PlantOutletNodeID(0), PlantMassFlowRate(0.0),
+                  PlantMassFlowRateMax(0.0), PlantMassFlowRateMaxWasAutoSized(false), AirInletNodeID(0), AirOutletNodeID(0), FuelSupplyID(0),
+                  DynamicsControlID(0), AvailabilitySchedID(0), CWLoopNum(0), CWLoopSideNum(0), CWBranchNum(0), CWCompNum(0)
+        {
+        }
+    };
+
+
     void SimMicroCHPGenerator(int const GeneratorType,          // type of Generator
                               std::string const &GeneratorName, // user specified name of Generator
                               int &GeneratorIndex,
@@ -141,6 +319,11 @@ namespace MicroCHPElectricGenerator {
                                      Real64 &ThermalPower,    // heat power
                                      Real64 &ThermalEnergy    // heat energy
     );
+
+    extern Array1D<MicroCHPDataStruct> MicroCHP;
+    extern Array1D<MicroCHPParamsNonNormalized> MicroCHPParamInput;
+
+    void clear_state();
 
 } // namespace MicroCHPElectricGenerator
